Make uploader use the TOML config file intrastructure, too.
